fix(clippy): complete Phase 1 low-risk warning fixes (15/24)
Automated and manual fixes for all low-risk clippy warnings:

Phase 1A - Automated fixes (cargo clippy --fix):
- Remove unnecessary cast in storage/persistence.rs
- Remove clone on Copy type in storage/tantivy.rs
- Remove needless borrows for format\! in tantivy.rs
- Simplify redundant closure in parsing/factory.rs
- Use or_default() instead of or_insert_with in resolver.rs
- Use first() instead of get(0) in simple.rs
- Use is_some_and() in walker.rs
- Remove unused enumerate index in clustering.rs
- Add Default implementations for ImportResolver, TantivyTransaction, FileTransaction
- Derive Default for RelationshipMetadata

Phase 1B/C - Manual fixes:
- Fix nested format\! calls in mcp/mod.rs
- Use strip_suffix() instead of manual string slicing in resolver.rs

Progress tracked in CLIPPY_FIX_PLAN.md for cross-session continuity.
9 warnings remain (medium/high risk) requiring careful review.

+++ src/storage/persistence.rs
@@ -30,7 +30,7 @@ impl IndexPersistence {
         
         metadata.update_counts(
             indexer.symbol_count() as u32,
-            indexer.file_count() as u32,
+            indexer.file_count(),
         );
         
         // Update metadata to reflect Tantivy
